requirements.txt
dbt-core>=1.7
dbt-postgres>=1.7
pdfplumber>=0.10
pandas>=2.0
psycopg2-binary>=2.9
requests>=2.31
python-dotenv>=1.0
supabase>=2.0
